% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster.fun.kkmeans}
\alias{cluster.fun.kkmeans}
\title{Wrapper function for kernal kmeans}
\usage{
cluster.fun.kkmeans(x, centers, ...)
}
\arguments{
\item{x}{the input data}

\item{centers}{the number of centers}

\item{...}{other parameters passing to \code{kernlab::kkmeans}}
}
\description{
Wrapper function for kernal kmeans
}
