% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster.fun.mlpack.old}
\alias{cluster.fun.mlpack.old}
\title{Kmeans Clustering from RcppMLPACK}
\usage{
cluster.fun.mlpack.old(x, centers, ...)
}
\arguments{
\item{x}{The input data for the clustering algorithm.}

\item{centers}{A number indicating the number of clustering centers.}

\item{...}{arguments for future use.}
}
\description{
The Kmeans algorithm from \code{RcppMLPACK}.
}

