% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster.predict.kkmeans}
\alias{cluster.predict.kkmeans}
\title{Predict function for kernel kmeans}
\usage{
cluster.predict.kkmeans(x, cluster.object)
}
\arguments{
\item{x}{The data to make prediction}

\item{cluster.object}{The result object from \code{kernlab::kkmeans}}
}
\description{
Predict function for kernel kmeans
}
