% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSVM.R
\name{csvmTransform}
\alias{csvmTransform}
\title{Data Transformation function for Clustered Support Vector Machine}
\usage{
csvmTransform(x, lambda, cluster.label, sparse = TRUE)
}
\arguments{
\item{x}{The data matrix, could be a \code{matrix} or \code{dgCMatrix} object.}

\item{lambda}{The parameter from the algorithm}

\item{cluster.label}{The clustering label starting from 1. Its length must equal to the number of rows in \code{x}}

\item{sparse}{Logical argument indicating whether the output should be a sparse matrix or not}
}
\description{
Transform a data matrix according to the kmeans clustering result based on 
  Gu, Quanquan, and Jiawei Han. "Clustered support vector machines."
}
