% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{eucliDist}
\alias{eucliDist}
\title{Euclidean Distance calculation}
\usage{
eucliDist(x, centers)
}
\arguments{
\item{x}{the data matrix}

\item{centers}{the matrix of centers}
}
\description{
Euclidean Distance calculation
}

