% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gater.R
\name{gater}
\alias{gater}
\title{Gater function for mixture SVMs}
\usage{
gater(x, y, S)
}
\arguments{
\item{x}{the nxp training data matrix. Could be a matrix or a sparse matrix object.}

\item{y}{a response vector for prediction tasks with one value for each of the n rows of \code{x}.
For classification, the values correspond to class labels and can be a 1xn matrix,
a simple vector or a factor. For regression, the values correspond to the values to predict,
and can be a 1xn matrix or a simple vector.}

\item{S}{the prediction matrix from experts}
}
\description{
Gater function for mixture SVMs
}

