% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gaterSVM.R
\name{gaterSVM}
\alias{gaterSVM}
\title{Mixture SVMs with gater function}
\usage{
gaterSVM(x, y, m, c = 1, max.iter)
}
\arguments{
\item{x}{the nxp training data matrix. Could be a matrix or a sparse matrix object.}

\item{y}{a response vector for prediction tasks with one value for each of the n rows of \code{x}.
For classification, the values correspond to class labels and can be a 1xn matrix,
a simple vector or a factor. For regression, the values correspond to the values to predict,
and can be a 1xn matrix or a simple vector.}

\item{m}{the number of experts}

\item{c}{a positive constant controlling the upper bound of
the number of samples in each subset.}

\item{max.iter}{the number of iterations}
}
\description{
Mixture SVMs with gater function
}

