% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{kern.predict}
\alias{kern.predict}
\title{Predict function for kernel kmeans}
\usage{
kern.predict(kkmeans.res, x)
}
\arguments{
\item{kkmeans.res}{The result object from \code{kernlab::kkmeans}}

\item{x}{The data to make prediction}
}
\description{
Predict function for kernel kmeans
}

