% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kmeans.predict}
\alias{kmeans.predict}
\title{Euclidean Distance based clustering prediction}
\usage{
kmeans.predict(x, cluster.object)
}
\arguments{
\item{x}{the data matrix}

\item{cluster.object}{the matrix of centers}
}
\description{
Euclidean Distance based clustering prediction
}
