% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{mnist}
\format{A list of six data objects \code{mnist38}, \code{mnist38.t}, \code{mnist49}, \code{mnist49.t},
  \code{mnistoe} and \code{mnistoe.t}.
  The first column is the target variable.}
\usage{
data(mnist)
}
\description{
Yann LeCun, L. Bottou, Y. Bengio, and P. Haffner. Gradient-based learning applied to document recognition.
  Proceedings of the IEEE, 86(11):2278-2324, November 1998.
}
\keyword{datasets}

