% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{oneclass.svm}
\alias{oneclass.svm}
\title{DIrectly assign alpha to svm}
\usage{
oneclass.svm(x, nclass, class_rank)
}
\arguments{
\item{x}{the data matrix}

\item{nclass}{the total number of classes}

\item{class_rank}{the rank of class that the input data matrix has}
}
\description{
DIrectly assign alpha to svm
}

