% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alphasvm.R
\name{plot.alphasvm}
\alias{plot.alphasvm}
\title{Plot alphasvm object}
\usage{
\method{plot}{alphasvm}(x, data, formula = NULL, fill = TRUE, grid = 50,
  slice = list(), symbolPalette = palette(), svSymbol = "x",
  dataSymbol = "o", ...)
}
\arguments{
\item{x}{An object of class \code{alphasvm}}

\item{data}{data to visualize. Should be the same used for fitting.}

\item{formula}{formula selecting the visualized two dimensions. Only needed if more than two input variables are used.}

\item{fill}{switch indicating whether a contour plot for the class regions should be added.}

\item{grid}{granularity for the contour plot.}

\item{slice}{a list of named values for the dimensions held constant (only needed if more than two variables are used).
The defaults for unspecified dimensions are 0 (for numeric variables) and the first level (for factors).
Factor levels can either be specified as factors or character vectors of length 1.}

\item{symbolPalette}{Color palette used for the class the data points and support vectors belong to.}

\item{svSymbol}{Symbol used for support vectors.}

\item{dataSymbol}{Symbol used for data points (other than support vectors).}

\item{...}{additional graphics parameters passed to \code{filled.contour} and \code{plot}.}
}
\description{
Plot alphasvm object
}

