% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterSVM.R
\docType{methods}
\name{predict,clusterSVM-method}
\alias{predict,clusterSVM-method}
\title{Predictions with Clustered Support Vector Machines}
\usage{
\S4method{predict}{clusterSVM}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class "clusterSVM", created by \code{clusterSVM}.}

\item{newdata}{An n x p matrix containing the new input data. Could be a matrix or a sparse matrix object.}

\item{...}{other parameters passing to \code{predict.LiblineaR}}
}
\description{
The function applies a model (classification or regression) produced by the
 \code{clusterSVM} function to every row of a data matrix and returns the model predictions.
}

