% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\name{predict.alphasvm}
\alias{predict.alphasvm}
\title{Prediction function for an alphasvm object}
\usage{
predict.alphasvm(object, newdata, decision.values = FALSE,
  probability = FALSE, ..., na.action = na.omit)
}
\description{
Prediction function for an alphasvm object
}

