% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSVM.R
\name{predict.clusterSVM}
\alias{predict.clusterSVM}
\title{Predictions with Clustered Support Vector Machines}
\usage{
\method{predict}{clusterSVM}(object, newdata = NULL, cluster.predict = NULL, ...)
}
\arguments{
\item{object}{Object of class "clusterSVM", created by \code{clusterSVM}.}

\item{newdata}{An n x p matrix containing the new input data. Could be a matrix or a sparse matrix object.}

\item{cluster.predict}{a function predict new labels on newdata.}

\item{...}{other parameters passing to \code{predict.LiblineaR}}
}
\description{
The function applies a model (classification) produced by the 
 \code{clusterSVM} function to every row of a data matrix and returns the model predictions.
}
