% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcSVM.R
\name{predict.dcSVM}
\alias{predict.dcSVM}
\title{Predictions with Divide-Conquer Support Vector Machines}
\usage{
\method{predict}{dcSVM}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class "dcSVM", created by \code{dcSVM}.}

\item{newdata}{An n x p matrix containing the new input data. Could be a matrix or a sparse matrix object.}

\item{...}{other parameters passing to \code{predict.svm}}
}
\description{
The function applies a model produced by the 
 \code{dcSVM} function to every row of a data matrix and returns the model predictions.
}
