% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gater.R
\name{predict.gater}
\alias{predict.gater}
\title{Predictions for Gater function}
\usage{
\method{predict}{gater}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class "gater", created by \code{gater}.}

\item{newdata}{An n x p matrix containing the new input data. Could be a matrix or a sparse matrix object.}

\item{...}{parameters for future usage.}
}
\description{
The function applies a model produced by the
 \code{gaterSVM} function to every row of a data matrix and returns the model predictions.
}

