% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaterSVM.R
\name{predict.gaterSVM}
\alias{predict.gaterSVM}
\title{Prediction for Gater SVM}
\usage{
\method{predict}{gaterSVM}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{gaterSVM}}

\item{newdata}{newdata An n x p matrix containing the new input data. 
Could be a matrix or a sparse matrix object.}

\item{...}{parameters for future usage.}
}
\description{
The function applies a model produced by the 
 \code{gaterSVM} function to every row of a data matrix and returns the model predictions.
}
