% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svm.R
\name{svm.default}
\alias{svm.default}
\title{export}
\usage{
\method{svm}{default}(x, y = NULL, scale = TRUE, type = NULL,
  kernel = "radial", degree = 3, gamma = if (is.vector(x)) 1 else
  1/ncol(x), coef0 = 0, cost = 1, nu = 0.5, class.weights = NULL,
  cachesize = 40, tolerance = 0.001, epsilon = 0.1, shrinking = TRUE,
  cross = 0, probability = FALSE, fitted = TRUE, alpha = NULL, ...,
  subset, na.action = na.omit)
}
\description{
export
}

