% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{usps}
\alias{usps}
\title{usps}
\format{Two data objects \code{usps} and \code{usps}.
  The first column is the target variable.}
\usage{
\code{data(usps)}
}
\description{
J. J. Hull. A database for handwritten text recognition research.
  IEEE Transactions on Pattern Analysis and Machine Intelligence, 16(5):550-554, May 1994.
}
\keyword{datasets}

