% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphasvm.R
\name{write.alphasvm}
\alias{write.alphasvm}
\title{Write alphasvm object}
\usage{
\method{write}{alphasvm}(
  object,
  svm.file = "Rdata.svm",
  scale.file = "Rdata.scale",
  yscale.file = "Rdata.yscale"
)
}
\arguments{
\item{object}{Object of class "alphasvm", created by \code{alphasvm}.}

\item{svm.file}{filename to export the \code{alphasvm} object to.}

\item{scale.file}{filename to export the scaling data of the explanatory variables to.}

\item{yscale.file}{filename to export the scaling data of the dependent variable to, if any.}
}
\description{
Write alphasvm object
}
