% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ijcnn1}
\alias{ijcnn1}
\title{ijcnn1}
\format{A list of two data objects \code{ijcnn1} and \code{ijcnn1.t}.
  The first column is the target variable.}
\usage{
data(ijcnn1)
}
\description{
Danil Prokhorov. IJCNN 2001 neural network competition.
  Slide presentation in IJCNN'01, Ford Research Laboratory, 2001.
}
\keyword{datasets}

