% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{svmguide1}
\alias{svmguide1}
\title{svmguide1}
\format{Two data objects \code{svmguide1} and \code{svmguide1.t}.
   The first column is the target variable.}
\usage{
\code{data(svmguide1)}
}
\description{
An astroparticle application from Jan Conrad of Uppsala University, Sweden.
}
\keyword{datasets}

